Add proper vertex buffer
This change introduces the use of a proper vertex buffer, meaning
the use of a Default type heap to allocate the buffer in. Previously
vertices were stored in an Upload type heap. An upload heap can be
written to by the CPU and red by the GPU, so it's easier to use
but introduces a significant performance cost. This method is not
recommended. Now, an upload heap is used to transfer vertex data from
the CPU to GPU, and then a command is issued that copies the contents
of the Upload Buffer to the much more performant Defalut Buffer, which
is used by the pipeline when rendering.

Additionally, include of the dxguid.lib was added to Linker Inputs,
as the lack of it started causing linker errors with IID_D3D12Device.