Change the RTVs (Render Target Views) to use sRGB
This change ensures correct color space transformations on the rendered
image. As the image is rendered in linerar color space, it needs to be
transormed into the sRGB space (gamma corection) to be displayed correctly.
Incorrect color space at the output results in darkened colors, loss
of information in dark areas and unrealistic visual effect.

This could also be done in shaders, by raising all final color values
to the power of the inverse gamma. This method ensures that the render
targets will move into sRGB space when rendering finishes.