Add texture loading - part 1
This change adds texture loading and textures the pyramids with them.
However, the program does not work correctly, as some faces do not
receive the texture and GPU validation layers spew errors.

In the next part the issues will be addressed by changing the engine
to use a single heap for all CBV/SRV/UAV resources. The wvp matrix
buffer will also be moved to the root signature as a root constant.

The root cause of the issue is that only one descriptor heap can be
set at a time within one command list. Currently in the application
i'm trying to set two, by setting one after the other. This is not
possible and results in the second heap overriding the selection
of the first one.