Fix resource heap binding
The code has been modified to use only one descriptor heap for CBV, SRV and UAV.
They need to be allocated and accesed via proper offsets into the heap.

This got rid of the validation layer lamentations, but curiously did
not fix the back half of the pyramids not receiving a texture.
Will need to further investigate that.