Add initial Mesh and Buffer abstractions
This commit:
1. Refactors existing code regarding buffers with the use of
the new BufferMemoryManager. In effect VoyagerEngine contains
less code and buffer related code is in one place. Also,
the BufferMemoryManager allows for some automation of the
process of allocation and filling buffers, as it will automatically
submit the commandList containing buffer oerations once it goes
out of scope. So one can bunch up FillBuffer operations in a
scope and then they will be automatically submitted to the GPU.
2. Adds a Mesh class, which is a basic abstracion of a vertex+index
buffer. It can be created from two vectors or from a file name (.obj).
It has two public methods operating on a commandList, used for rendering.
3. Additionally some utility functions and a DXContext were created,
latter of which will aggregate all things DX-related that need to be used
all over the codebase (like the device).