Add abstraction for textures
The Texture class can load a texture from file and create
an apropriate DX12 Resource. It also provides a method for
creating a view of the Resource in a given heap (presumably
an SRV heap/general shader access heap).

Hopefully this will offload some manual work from us, the developers.