Add material abstracions
The Material class defines a base material with one constant buffer
and one root signature constant. It is an abstraction of the
PSO and Root Signature DX components. It also loads shaders based
on name. It can be used as a base class, and inheriting classes
can override some of the methods of Material, mainly Root Signature
creation, to compose other specific materials (this is done using the
template method design pattern). Exactly that is done by the
DefaultTexturedMaterial. It adds to the Root Signature a table with
an SRV for a texture. Both materials are used during rendering.