Add a simple diffuse material
This change adds a new material that receives lighting.
It uses the most simple diffuse model. The shaders use
a light position taken from a constant buffer for these
calculations. The shading is done per-pixel.

Additionally, a bug in the vertex input layout was fixed,
which caused the normals to be read incorrectly, cutting off
some data.