Fix Material class
Change to wireframe rasterization can be done there for debugging
since changing a material and binding correct resources in correct
slots is still somewhat a hassle.

Such change however should not be commited.
Same goes for any debug changes that were made during development
of a feature but are not actually parts of the feature!