Fix flicker issue
Some meshes were flickering, because of synchronization issue.

Update didn't wait for the previous frame to render, so when the cpu
updated and submitted the 2 other frames in the framebuffer it would
start updating the data that the GPU has not finished working with.

This was fixed by moving the wait from the beginning of the render
method to the beginning of the update method. Now the CPU will wait
untill given framebuffers resources are avialable and the modify them.